refactor: move `fetch` test functions to test utils
