+++ tsconfig.json
@@ -1,7 +1,8 @@
 {
 	"exclude": ["out", "node_modules", ".vscode-test", "test/fixture"],
-  "extends": "@tsconfig/node18",
+  "extends": "@tsconfig/node18/tsconfig.json",
   "compilerOptions": {
+    "lib": ["es2023", "ESNext.Disposable"],
     "outDir": "out",
     "sourceMap": true,
     "rootDir": ".",
